api/pyproject.toml
[tool.poetry]
name = "syllabus-ai-api"
version = "0.1.0"
description = ""
authors = ["Shehab Shalan <shehab.shalan@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
fastapi-camelcase = "^2.0.0"
pydantic = "^2.9.2"
openai = "^1.50.1"
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
diskcache = "^5.6.3"
python-jose = "3.3.0"
google-auth = "2.25.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
